Fix specter bug
