+++ src/cryptoadvance/specter/devices/coldcard.py
@@ -1,6 +1,5 @@
-from .hwi_device import HWIDevice
-from ..serializations import PSBT
-from ..helpers import decode_base58, get_xpub_fingerprint, hash160
+from .sd_card_device import SDCardDevice
+from ..helpers import get_xpub_fingerprint
 
 
 CC_TYPES = {
@@ -8,30 +7,14 @@ CC_TYPES = {
     'p2sh-segwit': 'P2WSH-P2SH',
     'bech32': 'P2WSH'
 }
-class ColdCard(HWIDevice):
+class ColdCard(SDCardDevice):
     def __init__(self, name, alias, device_type, keys, fullpath, manager):
-        HWIDevice.__init__(self, name, alias, 'coldcard', keys, fullpath, manager)
+        SDCardDevice.__init__(self, name, alias, 'coldcard', keys, fullpath, manager)
         self.sd_card_support = True
         self.qr_code_support = False
 
     def create_psbts(self, base64_psbt, wallet):
-        psbts = HWIDevice.create_psbts(self, base64_psbt, wallet)
-        sdcard_psbt = PSBT()
-        sdcard_psbt.deserialize(base64_psbt)
-        if len(wallet.keys) > 1:
-            for k in wallet.keys:
-                key = b'\x01' + decode_base58(k.xpub)
-                if k.fingerprint != '':
-                    fingerprint = bytes.fromhex(k.fingerprint)
-                else:
-                    fingerprint = _get_xpub_fingerprint(k.xpub)
-                if k.derivation != '':
-                    der = _der_to_bytes(k.derivation)
-                else:
-                    der = b''
-                value = fingerprint + der
-                sdcard_psbt.unknown[key] = value
-        psbts['sdcard'] = sdcard_psbt.serialize()
+        psbts = SDCardDevice.create_psbts(self, base64_psbt, wallet)
         return psbts
 
     def get_wallet_file(self, wallet):
@@ -66,25 +49,3 @@ Format: {}
                 fingerprint = get_xpub_fingerprint(k.xpub).hex()
             cc_file += "{}: {}\n".format(fingerprint.upper(), k.xpub)
         return cc_file
-
-def _get_xpub_fingerprint(xpub):
-    b = decode_base58(xpub)
-    return hash160(b[-33:])[:4]
-
-def _der_to_bytes(derivation):
-    items = derivation.split("/")
-    if len(items) == 0:
-        return b''
-    if items[0] == 'm':
-        items = items[1:]
-    if items[-1] == '':
-        items = items[:-1]
-    res = b''
-    for item in items:
-        index = 0
-        if item[-1] == 'h' or item[-1] == "'":
-            index += 0x80000000
-            item = item[:-1]
-        index += int(item)
-        res += index.to_bytes(4,'big')
-    return res
